You get a truth table representing a Boolean function and must return its Canonical normal form (CDNF) as a string, using the following formatting/rules:
Variable names are always in order: "x1", "x2", ..., "xn".
The number of logical variables is between 1 and 5.
Output only terms where the result is true.
If the variable xi has the value "false" in the term, then it is written "~xi" in the output, otherwise "xi".
It is guaranteed that all truth tables are correctly built.